+++ jhotdraw8/src/main/java/org.jhotdraw8.draw/org/jhotdraw8/draw/handle/PathIterableOutlineHandle.java
@@ -61,7 +61,7 @@ public class PathIterableOutlineHandle extends AbstractHandle {
         CssColor color = view.getEditor().getHandleColor();
         path1.setStroke(Color.WHITE);
         path2.setStroke(Paintable.getPaint(color));
-        int strokeWidth = view.getEditor().getHandleStrokeWidth();
+        double strokeWidth = view.getEditor().getHandleStrokeWidth();
         path1.setStrokeWidth(strokeWidth + 2);
         path2.setStrokeWidth(strokeWidth);
         return node;
